pyproject.toml
[project]
name = "tmw-archive"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "dropbox>=12.0.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "laion-clap>=1.1.5",
    "librosa>=0.11.0",
    "nnaudio>=0.3.4",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "sqlalchemy>=2.0.48",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "umap-learn>=0.5.11",
    "uvicorn[standard]>=0.42.0",
]
